moved data/index -> data/dynamodb/index
