+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import { NextPage } from "next";
 import Image from "next/image";
 import { use } from "react";
 import favicon from "../../public/favicon.png";
-import { getChannels, getLatestVideos } from "../data";
+import { getChannels, getLatestVideos } from "../data/dynamodb";
 import styles from "./page.module.css";
 
 dayjs.extend(relativeTime);
