+++ src/data/dynamodb/index.ts
